fix: Forget about top level `tsconfig`
And create individual `tsconfig` files for each project.

This seems to work and deals with some of my concerns from
#2. We could potentially generate these `tsconfig` files
to make sure they reflected the actual state of stuff in
our monorepo and included all the correct dependencies.

I guess the question is whether it would have a disasterous
effect on performance / speed (especially in our editors).